+++ src/user/user_crypt.c
@@ -1,0 +1,6 @@
+#include "crypt.h"
+
+char* crypt(const char* key, const char* salt) {
+    (void)salt;
+    return (char*)key;
+}