+++ src/user/user_time.c
@@ -1,52 +1,10 @@
 #include "user_lib.h"
-#include "dirent.h"
-#include "crypt.h"
-#include "string.h"
 #include "time.h"
-#include "errno.h"
 
 static const char* s_months[] = {
     "Jan", "Feb", "Mar", "Apr", "May", "Jun",
     "Jul", "Aug", "Sep", "Oct", "Nov", "Dec"
 };
-
-#define FTP_COMPAT_DIR_POOL_SIZE 4
-
-static DIR s_dir_pool[FTP_COMPAT_DIR_POOL_SIZE];
-static unsigned char s_dir_pool_used[FTP_COMPAT_DIR_POOL_SIZE];
-
-char* strtok_r(char* s, const char* delim, char** saveptr) {
-    char* cur;
-    char* end;
-
-    if (s) {
-        cur = s;
-    } else if (saveptr && *saveptr) {
-        cur = *saveptr;
-    } else {
-        return 0;
-    }
-
-    cur += strspn(cur, delim);
-    if (*cur == '\0') {
-        if (saveptr) *saveptr = 0;
-        return 0;
-    }
-
-    end = cur + strcspn(cur, delim);
-    if (*end != '\0') {
-        *end = '\0';
-        if (saveptr) *saveptr = end + 1;
-    } else if (saveptr) {
-        *saveptr = 0;
-    }
-    return cur;
-}
-
-char* crypt(const char* key, const char* salt) {
-    (void)salt;
-    return (char*)key;
-}
 
 struct tm* gmtime_r(const time_t* timep, struct tm* result) {
     time_t v = timep ? *timep : time(0);
@@ -132,75 +90,4 @@ size_t strftime(char* s, size_t max, const char* format, const struct tm* tm) {
     }
     s[pos] = '\0';
     return pos;
-}
-
-DIR* opendir(const char* path) {
-    const char* use_path = (!path || !*path) ? "/" : path;
-    uint32_t size = 0;
-    int is_dir = 0;
-
-    int sr = sys_stat(use_path, &size, &is_dir);
-    if (sr < 0) {
-        errno = -sr;
-        return 0;
-    }
-    if (!is_dir) {
-        errno = ENOTDIR;
-        return 0;
-    }
-
-    for (unsigned int i = 0; i < FTP_COMPAT_DIR_POOL_SIZE; i++) {
-        if (!s_dir_pool_used[i]) {
-            DIR* d = &s_dir_pool[i];
-            s_dir_pool_used[i] = 1;
-            memset(d, 0, sizeof(*d));
-            strncpy(d->path, use_path, sizeof(d->path) - 1u);
-            d->index = 0;
-            return d;
-        }
-    }
-
-    {
-        DIR* d = (DIR*)malloc(sizeof(DIR));
-        if (!d) {
-            u_puts("ftp_compat: opendir alloc failed\n");
-            errno = ENOMEM;
-            return 0;
-        }
-        memset(d, 0, sizeof(*d));
-        strncpy(d->path, use_path, sizeof(d->path) - 1u);
-        d->index = 0;
-        return d;
-    }
-}
-
-struct dirent* readdir(DIR* dirp) {
-    int rc;
-    if (!dirp) {
-        errno = EBADF;
-        return 0;
-    }
-    rc = sys_dirlist(dirp->path, dirp->index, &dirp->current);
-    if (rc < 0) {
-        errno = -rc;
-        return 0;
-    }
-    if (rc == 0) return 0;
-    dirp->index++;
-    return &dirp->current;
-}
-
-int closedir(DIR* dirp) {
-    if (!dirp) {
-        errno = EBADF;
-        return -1;
-    }
-    for (unsigned int i = 0; i < FTP_COMPAT_DIR_POOL_SIZE; i++) {
-        if (dirp == &s_dir_pool[i]) {
-            s_dir_pool_used[i] = 0;
-            return 0;
-        }
-    }
-    free(dirp);
-    return 0;
 }